+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.20.0] - 2025-02-25
+
 ### Added
 
 - `installations` export now sets the `giantswarm.io/access-docs-markdown` annotation if extra access docs are provided in `docs/access.md` in the installation repository.
@@ -323,7 +325,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial code
 
-[Unreleased]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.19.0...HEAD
+[Unreleased]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.20.0...HEAD
+[0.20.0]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.19.0...v0.20.0
 [0.19.0]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.18.1...v0.19.0
 [0.18.1]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.18.0...v0.18.1
 [0.18.0]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.17.0...v0.18.0